requirements.txt
psycopg[binary]>=3.0.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
flask>=2.0.0
